+++ pkg/npm/plugin_impl.go
@@ -5,6 +5,13 @@
 
 package npm
 
+import (
+	"encoding/base64"
+	"fmt"
+	"net/url"
+	"strings"
+)
+
 // Settings for the Plugin.
 type Settings struct {
 	Username              string
@@ -18,6 +25,9 @@ type Settings struct {
 	Access                string
 }
 
+// globalRegistry defines the default NPM registry.
+const globalRegistry = "https://registry.npmjs.org"
+
 func (p *pluginImpl) Validate() error {
 	// Validate the Config and return an error if there are issues.
 	return nil
@@ -26,4 +36,27 @@ func (p *pluginImpl) Validate() error {
 func (p *pluginImpl) Exec() error {
 	// Implementation of the plugin.
 	return nil
+}
+
+// npmrcContentsUsernamePassword creates the contents from a username and
+// password
+func npmrcContentsUsernamePassword(config Settings) string {
+	// get the base64 encoded string
+	authString := fmt.Sprintf("%s:%s", config.Username, config.Password)
+	encoded := base64.StdEncoding.EncodeToString([]byte(authString))
+
+	// create the file contents
+	return fmt.Sprintf("_auth = %s\nemail = %s", encoded, config.Email)
+}
+
+/// Writes npmrc contents when using a token
+func npmrcContentsToken(config Settings) string {
+	registry, _ := url.Parse(config.Registry)
+	registry.Scheme = "" // Reset the scheme to empty. This makes it so we will get a protocol relative URL.
+	registryString := registry.String()
+
+	if !strings.HasSuffix(registryString, "/") {
+		registryString = registryString + "/"
+	}
+	return fmt.Sprintf("%s:_authToken=%s", registryString, config.Token)
 }
