+++ pkg/npm/plugin_test.go
@@ -7,6 +7,42 @@ package npm
 
 import "testing"
 
-func TestPlugin(t *testing.T) {
-	t.Skip()
+func TestTokenRCContents(t *testing.T) {
+	settings := Settings{
+		Registry: "https://npm.someorg.com/",
+		Token:    "token",
+	}
+	actual := npmrcContentsToken(settings)
+	expected := "//npm.someorg.com/:_authToken=token"
+	if actual != expected {
+		t.Errorf("Unexpected token settings (Got: %s, Expected: %s)", actual, expected)
+	}
+
+	settings.Registry = "https://npm.someorg.com/with/path/"
+	actual = npmrcContentsToken(settings)
+	expected = "//npm.someorg.com/with/path/:_authToken=token"
+	if actual != expected {
+		t.Errorf("Unexpected token settings (Got: %s, Expected: %s)", actual, expected)
+	}
+
+	settings.Registry = globalRegistry
+	actual = npmrcContentsToken(settings)
+	expected = "//registry.npmjs.org/:_authToken=token"
+	if actual != expected {
+		t.Errorf("Unexpected token settings (Got: %s, Expected: %s)", actual, expected)
+	}
+
+	settings.Registry = "https://npm.someorg.com"
+	actual = npmrcContentsToken(settings)
+	expected = "//npm.someorg.com/:_authToken=token"
+	if actual != expected {
+		t.Errorf("Unexpected token settings (Got: %s, Expected: %s)", actual, expected)
+	}
+
+	settings.Registry = "https://npm.someorg.com/with/path"
+	actual = npmrcContentsToken(settings)
+	expected = "//npm.someorg.com/with/path/:_authToken=token"
+	if actual != expected {
+		t.Errorf("Unexpected token settings (Got: %s, Expected: %s)", actual, expected)
+	}
 }
